move callback classes to util (also rename self.write -> self.callback)
